% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{formula.bbl}
\alias{formula.bbl}
\title{Formula in BBL Fitting}
\usage{
\method{formula}{bbl}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bbl}}

\item{...}{Other arguments}
}
\value{
Formula object
}
\description{
Returns the formula used in BBL fit
}
\examples{
titanic <- as.data.frame(Titanic)
fit <- bbl(Survived ~ Class + Sex + Age, data=titanic, weights=titanic$Freq)
formula(fit)
}
