% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{freq2raw}
\alias{freq2raw}
\title{Convert Frequency Table into Raw Data}
\usage{
freq2raw(fdata, Freq)
}
\arguments{
\item{fdata}{Data frame with factors in columns and a frequency column}

\item{Freq}{Integer column index or column name of frequency}
}
\value{
Raw data frame with one row per instances
}
\description{
Data with unique rows and a frequency column is converted into
data with duplicate rows.
}
\details{
The ouput data frame can be used as input to \code{\link{bbl}}.
}
\examples{
Titanic
x <- as.data.frame(Titanic)
head(x)
titanic <- freq2raw(x, Freq='Freq')
head(titanic)
}
