% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.bbl}
\alias{logLik.bbl}
\title{Log likelihood for bbl object}
\usage{
\method{logLik}{bbl}(object, ...)
}
\arguments{
\item{object}{Object of class \code{bbl}}

\item{...}{Other arguments to methods}
}
\value{
Log likelihood value
}
\description{
Compute log likelihood from a fitted \code{bbl} object
}
\details{
This method will use inferred parameters from calls to \code{bbl} and data
compute the log likelihood.
}
