% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{model.frame.bbl}
\alias{model.frame.bbl}
\title{Model Frame for BBL}
\usage{
\method{model.frame}{bbl}(formula, ...)
}
\arguments{
\item{formula}{Object of class \code{bbl}}

\item{...}{Other arguments}
}
\value{
Data frame used for fitting
}
\description{
Returns the model frame used in BBL fit
}
\examples{
titanic <- as.data.frame(Titanic)
fit <- bbl(Survived ~ Class + Sex + Age, data=titanic[,1:4], weights=titanic$Freq)
head(model.frame(fit))
}
