% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs.bbl}
\alias{nobs.bbl}
\title{Number of Observations in BBL Fit}
\usage{
\method{nobs}{bbl}(object, ...)
}
\arguments{
\item{object}{Object of class \code{bbl}}

\item{...}{Other arguments}
}
\value{
An integer of number of observations
}
\description{
Returns the number of observations from a BBL fit
}
\examples{
titanic <- as.data.frame(Titanic)
fit <- bbl(Survived ~ Class + Sex + Age, data=titanic[,1:4], weights=titanic$Freq)
nobs(fit)
}
