% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.bbl}
\alias{plot.bbl}
\title{Plot bbl object}
\usage{
\method{plot}{bbl}(x, layout = NULL, hcol = NULL, Jcol = NULL, npal = 100, ...)
}
\arguments{
\item{x}{Object of class \code{bbl}}

\item{layout}{Matrix of layouts for arrangment of linear and interaction
parameters. If \code{NULL}, the top half will be used for linear
parameter barplot and bottom half will be divided into interaction
heatmaps for each response group.}

\item{hcol}{Color for linear barplots. Grayscale if \code{NULL}.}

\item{Jcol}{Color for interaction heatmaps. Default (\code{NULL}) is
\code{RdBu} from \code{RColorBrewer}.}

\item{npal}{Number of color scales.}

\item{...}{Other graphical parameters for \code{\link{plot}}.}
}
\description{
Visualize bias and interaction parameters
}
\details{
This method displays a barplot of bias parameters and heatmaps
(one per response group) of interaction parameters. All parameters are
offset by the pooled values (single group inference) unless missing.
}
