% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.cv.bbl}
\alias{plot.cv.bbl}
\title{Plot Cross-validation Outcome}
\usage{
\method{plot}{cv.bbl}(x, type = "b", log = "x", ...)
}
\arguments{
\item{x}{Object of class \code{cv.bbl} from a call to
\code{\link{crossVal}}}

\item{type}{Symbol type in \code{\link{plot}}, present here to set default.}

\item{log}{Log scale argument to \code{\link{plot}}.}

\item{...}{Other arguments to \code{\link{plot}}.}
}
\description{
Plot cross-validation score as a function of regularization parameter
}
\details{
This function will plot accuracy score as a function of regularization parameter
from a call to \code{\link{crossVal}}.
}
