% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.bbl}
\alias{print.bbl}
\title{Print Boltzmann Bayes Learning Fits}
\usage{
\method{print}{bbl}(x, showcoeff = TRUE, maxcoeff = 3L, ...)
}
\arguments{
\item{x}{An object of class \code{bbl}, usually dervied from a call to
\code{\link{bbl}}.}

\item{showcoeff}{Display first few fit coefficients}

\item{maxcoeff}{Maximum number of coefficients to display}

\item{...}{Further arguments passed to or from other methods}
}
\description{
This method displays model structure and first elements of coefficients
}
\details{
Displays the call to \code{\link{bbl}}, response variable and its levels,
predictors and their levels, and the first few fit coefficients.
}
