% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cv.bbl}
\alias{print.cv.bbl}
\title{Display Cross-validation Result}
\usage{
\method{print}{cv.bbl}(x, ...)
}
\arguments{
\item{x}{Object of class \code{cv.bbl}}

\item{...}{Other arguments to methods}
}
\description{
Print cross-validation optimal result and data frame
}
\details{
This method prints \code{\link{crossVal}} object with the optimal
regularization condition and maximum accuracy score on top and
the entire score profile as a data frame below.
}
