% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.bbl}
\alias{print.summary.bbl}
\title{Print Summary of Boltzmann Bayes Learning}
\usage{
\method{print}{summary.bbl}(x, ...)
}
\arguments{
\item{x}{Object of class \code{summary.bbl}}

\item{...}{Other arguments to methods}
}
\description{
This method prints the summary of \code{bbl} object
}
\details{
The naive Bayes summary of \code{summary.bbl} object is displayed.
}
