% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{weights.bbl}
\alias{weights.bbl}
\title{Weights in BBL Fit}
\usage{
\method{weights}{bbl}(object, ...)
}
\arguments{
\item{object}{Object of class \code{bbl}.}

\item{...}{Other arguments}
}
\value{
Vector of weights for each instance
}
\description{
This method returns weights used in BBL fit.
}
\details{
Note that weithts are integral
frequency values specifying repeat number of each instance in \code{bbl}.
If no weights were used (default of 1s), \code{NULL} is returned.
}
